fix: 근태정정로직, 휴가신청로직 수정 - 직원 신청내역 조회 페이지 추가 - dto위치 수정

+++ backend/src/main/java/com/kh/ct/domain/emp/entity/Emp.java
@@ -64,7 +64,7 @@ public class Emp extends BaseTimeEntity {
 
     private LocalDateTime endDate;
 
-    private Float leaveCount;
+    private Double leaveCount;
 
     @Column(nullable = false, length = 50, unique = true)
     private String empNo;
